fix(mcp): address eschutho review — structured errors, shared util, alias tests
- Move _sanitize_validation_error to shared utils/error_sanitization.py
- Use structured error dict in update_chart_preview (was plain string)
- Fix before-validator alias check to use key presence not truthiness
- Add performance block to update_chart parse error response
- Fix inconsistent wording in generate_explore_link error message
- Add 4 alias tests: title→dashboard_title, name→chart_name,
  XSS via alias rejection, first-alias-wins priority
