+++ superset/mcp_service/chart/tool/update_chart.py
@@ -327,7 +327,7 @@ async def update_chart(  # noqa: C901
             try:
                 parsed_config = parse_chart_config(request.config)
             except (ValueError, TypeError) as e:
-                from superset.mcp_service.chart.validation.pipeline import (
+                from superset.mcp_service.utils.error_sanitization import (
                     _sanitize_validation_error,
                 )
 
@@ -340,6 +340,11 @@ async def update_chart(  # noqa: C901
                             "message": f"Invalid chart configuration: {sanitized}",
                             "details": sanitized,
                             "error_code": "INVALID_CHART_CONFIG",
+                        },
+                        "performance": {
+                            "query_duration_ms": int((time.time() - start_time) * 1000),
+                            "cache_status": "error",
+                            "optimization_suggestions": [],
                         },
                         "success": False,
                         "schema_version": "2.0",
